Can remove friends

+++ routes/web.php
@@ -47,7 +47,7 @@ Route::prefix('/api')->middleware(['auth'])->group(function () {
         Route::get('/pseudo/{pseudo}', [UserController::class, 'searchByUserPseudo']);
         Route::post('/friends/send', [FriendController::class, 'sendFriendRequest']);
         Route::patch('/friends/action', [FriendController::class, 'handleFriendRequest']);
-        Route::delete('/friends/remove', [FriendController::class, 'removeFriend']);
+        Route::delete('/friends/remove/{user_id}', [FriendController::class, 'removeFriend']);
     });
 
     Route::prefix('/ors')->middleware(['set.ors.api.key'])->group(function () {
